refactor: clean up and co-locate item conversion code
