+++ src/actions/editDashboard.js
@@ -15,13 +15,8 @@ import { sGetEditDashboardRoot } from '../reducers/editDashboard';
 import { updateDashboard, postDashboard } from '../api/editDashboard';
 import { tSetSelectedDashboardById } from '../actions/selected';
 import { NEW_ITEM_SHAPE } from '../components/ItemGrid/gridUtil';
-import {
-    itemTypeMap,
-    isSpacerType,
-    TEXT,
-    emptyTextItemContent,
-    isTextType,
-} from '../modules/itemTypes';
+import { itemTypeMap } from '../modules/itemTypes';
+import { convertUiItemsToBackend } from '../modules/uiBackendItemConverter';
 
 const onError = error => {
     console.log('Error (Saving Dashboard): ', error);
@@ -96,23 +91,9 @@ export const acRemoveDashboardItem = value => ({
 export const tSaveDashboard = () => async (dispatch, getState) => {
     const dashboard = sGetEditDashboardRoot(getState());
 
-    const dashboardItems = dashboard.dashboardItems.map(item => {
-        const text = isTextType(item)
-            ? item.text || emptyTextItemContent
-            : null;
-
-        const type = isSpacerType(item) ? TEXT : item.type;
-
-        return {
-            ...item,
-            ...(text ? { text } : {}),
-            type,
-        };
-    });
-
     const dashboardToSave = {
         ...dashboard,
-        dashboardItems,
+        dashboardItems: convertUiItemsToBackend(dashboard.dashboardItems),
     };
 
     try {
