+++ src/modules/itemTypes.js
@@ -32,11 +32,6 @@ const DOMAIN_TYPE_TRACKER = 'TRACKER';
 
 // Dashboard helpers
 export const spacerContent = 'SPACER_ITEM_FOR_DASHBOARD_LAYOUT_CONVENIENCE';
-export const emptyTextItemContent = 'TEXT_ITEM_WITH_NO_CONTENT';
-export const isSpacerType = item =>
-    item.type === TEXT && item.text === spacerContent;
-export const isTextType = item =>
-    item.type === TEXT && item.text !== spacerContent;
 export const isVisualizationType = item =>
     !!itemTypeMap[item.type].isVisualizationType;
 export const hasMapView = itemType =>
